move constants into platform
