+++ src/webviews/extension-side/variablesView/variableView.ts
@@ -4,7 +4,7 @@
 import '../../../platform/common/extensions';
 import { Uri, WebviewView as vscodeWebviewView } from 'vscode';
 import { joinPath } from '../../../platform/vscode-path/resources';
-import { captureTelemetry, sendTelemetryEvent } from '../../../telemetry';
+import { captureTelemetry, sendTelemetryEvent, Telemetry } from '../../../telemetry';
 import { INotebookWatcher, IVariableViewPanelMapping } from './types';
 import { VariableViewMessageListener } from './variableViewMessageListener';
 import { InteractiveWindowMessages, IShowDataViewer } from '../../../platform/messageTypes';
@@ -30,7 +30,6 @@ import {
     IExtensionContext
 } from '../../../platform/common/types';
 import * as localize from '../../../platform/common/utils/localize';
-import { Telemetry } from '../../webview-side/common/constants';
 import { DataViewerChecker } from '../dataviewer/dataViewerChecker';
 import { IJupyterVariableDataProviderFactory, IDataViewerFactory, IDataViewer } from '../dataviewer/types';
 import { WebviewViewHost } from '../../../platform/webviews/webviewViewHost';
