+++ src/webviews/webview-side/plot/mainPanel.tsx
@@ -18,7 +18,7 @@ import { createDeferred } from '../../../platform/common/utils/async';
 import { SharedMessages } from '../../../platform/messageTypes';
 import { IPlotViewerMapping, PlotViewerMessages } from '../../extension-side/plotting/types';
 import { IJupyterExtraSettings } from '../../../platform/webviews/types';
-import { RegExpValues } from '../common/constants';
+import { RegExpValues } from '../../../platform/common/constants';
 
 // Our css has to come after in order to override body styles
 export interface IMainPanelProps {
